feat: user_id 생성 함수 추가
